refactor spanner client
